requirements.txt
Django==2.1.4
django-picklefield==2.0
pytz==2018.7
